requirements.txt
Django==4.2
psycopg2-binary==2.9.6  # PostgreSQL adapter
google-api-python-client==2.85.0  # Google Calendar API
google-auth==2.15.0
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
requests==2.28.1  # Para la integración con ManyChat
djangorestframework==3.14.0  # Django REST Framework
gunicorn
openai